2-nd step of normination (base_flag)

+++ libft/libft.h
@@ -37,7 +37,7 @@ char				*ft_strcpy (char *dest, const char *src);
 char				*ft_strdup(const char *str);
 void				ft_bzero(void *s, size_t n);
 size_t				ft_strlcat(char *dest, const char *src, size_t size);
-size_t				ft_strlen(const char *str);
+int					ft_strlen(const char *str);
 char				*ft_strncat(char *dest, const char *str, size_t n);
 char				*ft_strncpy(char *dest, const char *src, size_t n);
 char				*ft_strrchr (const char *str, int ch);
